Fixed type inconsistencies

+++ benchmark/include/DynamicAdaptBenchmark.hpp
@@ -226,7 +226,7 @@ public:
         std::ofstream add_stream(addfile);
         time_t cur_time;
 
-        for(size_t stage = 0; stage < layers_to_create.size(); stage++) {
+        for(uint32_t stage = 0; stage < layers_to_create.size(); stage++) {
             time_t duration = ((uint64_t)durations[stage]) * 1000L * 1000L;   // Seconds to microseconds
             time_t start_time = get_timestamp();
             for(size_t i = 0; i < layers_to_create[stage].size(); i++) {
@@ -255,7 +255,7 @@ public:
         std::ofstream del_stream(delfile);
         time_t cur_time;
 
-        for(size_t stage = 0; stage < layers_to_delete.size(); stage++) {
+        for(uint32_t stage = 0; stage < layers_to_delete.size(); stage++) {
             time_t duration = ((uint64_t)durations[stage]) * 1000L * 1000L;   // Seconds to microseconds
             time_t start_time = get_timestamp();
             for(size_t i = 0; i < layers_to_delete[stage].size(); i++) {
